CRAYSAT-1343: Hack the build script for feature/CRAYSAT-1343 branches
Hack the build scripts, config files, and docker compose files to allow
us to build Eli's branches of the docs-sat repo for CRAYSAT-1343. This
is helping me to review the changes, so I can easily see how things
render in the HTML.

Test Description:
Ran `bin/build.sh sat` and `bin/dev.sh sat` and verified that I could
browse the 4 versions of the docs in my browser.