CRAYSAT-1343: Add SOURCE_SUBDIR parameter
This parameter allows the docs build to point to a subdirectory
of the documentation instead of the top-level. This allows you to
put in files like CONTRIBUTING.md, README.md (describing the repo rather
than the overall documentation), or other files that should not end
up in the rendered product.

This commit also fixes a typo in the SAT instructions.

(cherry picked from commit b151540a3162a2c4bb36bffaa07ed62a8ed1da4f)